requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-addanother==2.0.0
django-crispy-forms==1.6.1
django-crudbuilder==0.2.4
django-extensions==1.6.3
django-nose==1.4.4
django-redis==4.6.0
django-tables2==1.2.6
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
fake-factory==0.7.2
gunicorn==19.4.5
nose==1.3.7
Pillow==3.3.0
PyJWT==1.4.2
python-dateutil==2.6.0
redis==2.10.5
six==1.10.0
wheel==0.26.0
whitenoise==3.0
